chore: update

+++ dirt-example/dirt-example-ddd-uc/src/main/java/com/zk/uc/repository/PoiRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PoiRepository extends JpaRepository<Poi, Integer> {
+public interface PoiRepository extends JpaRepository<Poi, Long> {
 }
